Make vertex coordinate relative values use highp + Make UniformBlock for VisualRenderer
- Make texture coordinate relative codes use highp instead of mediump
- Seperate VertexShader's uniform blocks for VisualRenderer, so we can
  easly control them by SharedUniform in future.

+++ dali-scene3d/internal/graphics/shaders/shadow-map-shader.frag
@@ -1,3 +1,5 @@
+//@name shadow-map-shader.frag
+
 //@version 100
 
 UNIFORM_BLOCK FragBlock
@@ -8,8 +10,8 @@ UNIFORM_BLOCK FragBlock
   UNIFORM int uIsShadowCasting;
 };
 
-INPUT mediump vec2 vUV;
-INPUT lowp vec4 vColor;
+INPUT highp vec2 vUV;
+INPUT highp vec4 vColor;
 
 #ifdef THREE_TEX
 #ifdef BASECOLOR_TEX
