+++ dali-toolkit/internal/graphics/shaders/page-turn-book-spine-effect.frag
@@ -2,8 +2,9 @@
 
 //@version 100
 
-precision mediump float;
-INPUT mediump vec2 vTexCoord;
+precision highp float;
+
+INPUT highp vec2 vTexCoord;
 UNIFORM sampler2D sTexture;
 
 UNIFORM_BLOCK FragBlock
